STYLE: Move to Mesh directory
Move all code handling meshes to the Mesh directory. Future commits
plan to refactor the classes in this folder.
Remove useless lib/inteface directory.
