+++ CMake/ProjectConfig.cmake
@@ -4,11 +4,10 @@ set( APPLICATION_SRC_DIRS
   "${SRC_DIR}/local/Interface"
   "${SRC_DIR}/local/Math"
   "${SRC_DIR}/local/Math/3d"
-  "${SRC_DIR}/local/Perlin"
+  "${SRC_DIR}/local/Math/Perlin"
+  "${SRC_DIR}/local/Mesh"
   "${SRC_DIR}/local/ProceduralMap" )
 
 set( LIB_SRC_DIRS
   "${SRC_DIR}/lib/common"
-  "${SRC_DIR}/lib/interface"
-  "${SRC_DIR}/lib/mesh"
   "${SRC_DIR}/lib/opengl" )
