+++ src/local/Mesh/format/mesh_io_obj.cpp
@@ -16,9 +16,9 @@
 **    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include "mesh.hpp"
 #include "mesh_io_obj.hpp"
-#include "../../common/error_handling.hpp"
-#include "../../mesh/mesh.hpp"
+#include "error_handling.hpp"
 
 #include <sstream>
 #include <fstream>
