+++ src/local/Mesh/format/mesh_io_off.cpp
@@ -16,9 +16,9 @@
 **    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include "mesh.hpp"
 #include "mesh_io_off.hpp"
-#include "../../common/error_handling.hpp"
-#include "../../mesh/mesh.hpp"
+#include "error_handling.hpp"
 
 #include <iostream>
 #include <fstream>
